Replace travis with github actions (#169)
