+++ .nvmrc
@@ -1,0 +1,1 @@
+12.16.1
