+++ cypress.json
@@ -3,5 +3,6 @@
   "defaultCommandTimeout": 8000,
   "viewportWidth": 1100,
   "viewportHeight": 1000,
-  "video": false
+  "video": false,
+  "projectId": "ivst8i"
 }
