+++ cypress/integration/cases1.spec.ts
@@ -1,0 +1,12 @@
+import { exectueTestcase } from './executeCase'
+
+describe('Cases', () => {
+  ;[
+    'base',
+    'complete',
+    'completeDecimal',
+    'withPartner',
+    'withEmployment',
+    // 'withMortgage',
+  ].forEach(exectueTestcase)
+})
