+++ cypress/integration/cases2.spec.ts
@@ -1,0 +1,11 @@
+import { exectueTestcase } from './executeCase'
+
+describe('Cases', () => {
+  ;[
+    'withPension',
+    'withChildren',
+    'with2percent',
+    'withSpa',
+    'withBonus',
+  ].forEach(exectueTestcase)
+})
