+++ cypress/integration/cases3.spec.ts
@@ -1,0 +1,11 @@
+import { exectueTestcase } from './executeCase'
+
+describe('Cases', () => {
+  ;[
+    'withEmploymentBonus',
+    'withHighIncome',
+    'bugReport1',
+    'bugReport2',
+    'bugReport3',
+  ].forEach(exectueTestcase)
+})
