+++ cypress/integration/postponeCases.spec.ts
@@ -1,0 +1,5 @@
+import { executePostponeCase } from './executeCase'
+
+describe.skip('Postpone cases', () => {
+  ;['basic', 'foreignIncome'].forEach(executePostponeCase)
+})
