+++ package.json
@@ -9,13 +9,14 @@
     "dev:cy": "start-server-and-test dev http://localhost:3000/ cy:open",
     "build": "BUILD_TIMESTAMP=$(date +%s) next build",
     "start": "next start",
-    "test": "npm run test:noWatch -- --watch",
-    "test:noWatch": "jest",
-    "ci": "npm run lint && npm run test:noWatch -- --ci && WITH_DEBUG=true npm run build && npm run cy:ci",
+    "test:watch": "npm run test -- --watch",
+    "test": "jest",
+    "ci": "npm run lint && npm run test -- --ci && WITH_DEBUG=true npm run build && npm run cy:ci",
     "prettier": "prettier --write  './{src,__tests__,cypress}/**/*.{ts,tsx,js,css}'",
     "cy:open": "cypress open",
     "cy:run": "cypress run",
-    "cy:ci": "start-server-and-test start http://localhost:3000/ cy:run",
+    "cy:run:dashboard": "cypress run --record --parallel",
+    "cy:ci": "start-server-and-test start http://localhost:3000/ cy:run:dashboard",
     "lint": "eslint . --format pretty --ext .js,.jsx,.ts,.tsx --max-warnings=0",
     "check:ts": "tsc --noEmit"
   },
